Reestruturação dos módulos
Módulos da crate reorganizados, sendo divididos em:
- base/: módulos/tipos referentes a representação sem modelo definido;
- padrao/: módulos/tipos de uso comum a toda a crate;
- /: armazenamento dos módulos/tipos dos modelos já implementados;
