+++ nfe/src/base/mod.rs
@@ -1,0 +1,4 @@
+//! Tipos e módulos com implementações restritas
+//! a base da NF-e, sem distinções dos modelos
+
+pub mod nfe;
