+++ nfe/src/lib.rs
@@ -1,17 +1,13 @@
 //! nfe - Crate para acesso aos dados da Nota Fiscal Eletrônica
 
-mod ide;
-mod emit;
-mod dest;
+mod padrao;
 mod nfe;
-mod nfe_base;
+
+pub mod base;
 
 pub use crate::{
-    nfe_base::*,
     nfe::*,
-    emit::*,
-    dest::*,
-    ide::*
+    padrao::*
 };
 
 #[cfg(test)]
