+++ nfe/src/padrao/dest.rs
@@ -7,9 +7,8 @@ pub struct Destinatario {
 }
 
 impl Destinatario {
-
     /// Parse da seção <emit>
-    pub(crate) fn parse(xml: &Dom) -> Result<Option<Self>, String> {
+    pub(crate) fn parse(xml: &Dom) -> Result<Option<Destinatario>, String> {
 
         let mut t_dest = Dom::new(DomType::Tag);
         t_dest.set_tag(Tag::new("dest"));
